Attempt to fix webpack module resolution by renaming graphql.ts module.
May fix #8861, if my hunch is correct that webpack is resolving the
```ts
import { isType } from 'graphql'
```
as importing from `utilities/globals/graphql.ts` module itself, rather
than the `graphql` package, as suggested by the "(possible exports:
removeTemporaryGlobals)" part of the error message:
```
WARNING in ./node_modules/@apollo/client/utilities/globals/graphql.js 4:4-10
export 'isType' (imported as 'isType') was not found in 'graphql' (possible exports: removeTemporaryGlobals)
```
The only relevant module that exports `removeTemporaryGlobals` is
`utilities/globals/graphql.ts`, so renaming that module to something
that cannot be confused with the `graphql` package should either fix the
problem or at least shed more light on what's going on.
