+++ src/utilities/globals/fix-graphql.ts
